fix some location clang-tidy URL: https://qingluan.hozonauto.com/geelib/repoMergeDetail?repoId=algorithm/mapping&mergeId=200

+++ modules/location/coord_adapter/lib/defines.h
@@ -15,7 +15,7 @@ namespace loc {
 namespace ca {
 
 struct Params {
-  double dr_deque_capacity = 100.0;
+  uint32_t dr_deque_capacity = 100.0;
 };
 
 struct Node {
